delete button

+++ src/main/java/org/example/client/service/UpdateStatus.java
@@ -16,7 +16,7 @@ import jakarta.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="arg0" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/&gt;
+ *         &lt;element name="arg0" type="{http://www.w3.org/2001/XMLSchema}int"/&gt;
  *         &lt;element name="arg1" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
@@ -33,30 +33,22 @@ import jakarta.xml.bind.annotation.XmlType;
 })
 public class UpdateStatus {
 
-    protected Long arg0;
+    protected int arg0;
     protected String arg1;
 
     /**
      * Obtém o valor da propriedade arg0.
      * 
-     * @return
-     *     possible object is
-     *     {@link Long }
-     *     
      */
-    public Long getArg0() {
+    public int getArg0() {
         return arg0;
     }
 
     /**
      * Define o valor da propriedade arg0.
      * 
-     * @param value
-     *     allowed object is
-     *     {@link Long }
-     *     
      */
-    public void setArg0(Long value) {
+    public void setArg0(int value) {
         this.arg0 = value;
     }
